Building a skeleton for zfs. -Separating zfs and vhost into a separate package -Improving the qemu target structure to generate multiple VMs with zfs volume -Fix minor bugs -Adding options for zfs
