+++ autobench/autobench.go
@@ -18,7 +18,6 @@ type Options struct {
 	TargetFIODevice    string `short:"D" long:"targetdev" description:"[Optional] To specify block device as a target for FIO. Needs superuser rights (-u=root)."`
 	LocalFolderResults string `short:"f" long:"folder" description:"[Optional] A name of folder with tests results" default:"FIOTestsResults"`
 	LocalDirResults    string `short:"l" long:"localpath" description:"[Optional] Path to directory with test results"`
-	TargetDisk         string `long:"target-disk" description:"Physical device run the tests on "`
 }
 
 var opts Options
